Archives: fix size of members on 64bit linux/unix
Windows and Linux/Posix have different opinions on the length of
"unsigned long" and other C types on 64bit:  Windows compilers
keep 32bit and 64bit identical, while they differ on POSIX.

Work around this by using the explicitly sized types, and letting
the compiler sort it out.

+++ TheForceEngine/TFE_Archive/gobMemoryArchive.cpp
@@ -36,7 +36,7 @@ bool GobMemoryArchive::open(const u8* buffer, size_t size)
 	m_header   = (GobArchive::GOB_Header_t*)readBuffer;
 	readBuffer = m_buffer + m_header->MASTERX;
 
-	m_fileList.MASTERN = *((long*)readBuffer); readBuffer += sizeof(long);
+	m_fileList.MASTERN = *((u32*)readBuffer); readBuffer += sizeof(u32);
 	m_fileList.entries = (GobArchive::GOB_Entry_t*)(readBuffer);
 
 	m_archiveOpen = true;
